# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for getting order information",
        "title": "L0 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/orders/{orderID}": {
            "get": {
                "description": "Gets information about an order by its ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Receive an order by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "orderID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CombinedData"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CombinedData": {
            "description": "Information about the order and nested structures",
            "type": "object",
            "properties": {
                "delivery": {
                    "description": "delivery information",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Delivery"
                        }
                    ]
                },
                "items": {
                    "description": "items information",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Item"
                    }
                },
                "order": {
                    "description": "Main order information",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Order"
                        }
                    ]
                },
                "payment": {
                    "description": "payment information",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Payment"
                        }
                    ]
                }
            }
        },
        "models.Delivery": {
            "description": "delivery information",
            "type": "object",
            "required": [
                "address",
                "city",
                "email",
                "id",
                "name",
                "phone",
                "region",
                "zip"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "zip": {
                    "type": "string"
                }
            }
        },
        "models.Item": {
            "description": "Product information",
            "type": "object",
            "required": [
                "brand",
                "chrtID",
                "id",
                "name",
                "nmID",
                "orderUID",
                "price",
                "rid",
                "sale",
                "size",
                "status",
                "totalPrice",
                "trackNumber"
            ],
            "properties": {
                "brand": {
                    "type": "string"
                },
                "chrtID": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nmID": {
                    "type": "integer"
                },
                "orderUID": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "rid": {
                    "type": "string"
                },
                "sale": {
                    "type": "integer"
                },
                "size": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "totalPrice": {
                    "type": "integer"
                },
                "trackNumber": {
                    "type": "string"
                }
            }
        },
        "models.Order": {
            "description": "Main order information",
            "type": "object",
            "required": [
                "customerID",
                "dateCreated",
                "deliveryID",
                "deliveryService",
                "entry",
                "locale",
                "oofShard",
                "orderUID",
                "paymentID",
                "shardKey",
                "smID",
                "trackNumber"
            ],
            "properties": {
                "customerID": {
                    "type": "string"
                },
                "dateCreated": {
                    "type": "string"
                },
                "deliveryID": {
                    "type": "string"
                },
                "deliveryService": {
                    "type": "string"
                },
                "entry": {
                    "type": "string"
                },
                "internalSignature": {
                    "type": "string"
                },
                "locale": {
                    "type": "string"
                },
                "oofShard": {
                    "type": "string"
                },
                "orderUID": {
                    "type": "string"
                },
                "paymentID": {
                    "type": "string"
                },
                "shardKey": {
                    "type": "string"
                },
                "smID": {
                    "type": "integer"
                },
                "trackNumber": {
                    "type": "string"
                }
            }
        },
        "models.Payment": {
            "description": "payment information",
            "type": "object",
            "required": [
                "amount",
                "bank",
                "currency",
                "customFee",
                "deliveryCost",
                "goodsTotal",
                "id",
                "paymentDT",
                "provider",
                "transaction"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "bank": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "customFee": {
                    "type": "integer"
                },
                "deliveryCost": {
                    "type": "integer"
                },
                "goodsTotal": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "paymentDT": {
                    "type": "integer"
                },
                "provider": {
                    "type": "string"
                },
                "requestID": {
                    "type": "string"
                },
                "transaction": {
                    "type": "string"
                }
            }
        }
    }
}